Don't re-add kexts and the kernel each time we get a notification that kexts were newly added.
The Darwin userland dynamic loader provides lldb with a list of 
newly-added or newly-removed binaries but in the kernel case we
only know that something has changed.  DynamicLoaderDarwinKernel
really needs to maintain its own persistent list of kexts that 
it has been notified about (most importantly, it will not detect
kext unlods) but for now we'll at least avoid re-adding an already
present kext.

<rdar://problem/12658487>, <rdar://problem/12658487> 


git-svn-id: https://llvm.org/svn/llvm-project@169082 91177308-0d34-0410-b5e6-96231b3b80d8