Unbaseline, but mark as untested, gdb.dwarf2/dw2-skip-prologue.
This test relies on ordering of inline asm relative to functions to produce
portable source with hardcoded DWARF in a side-by-side asm file. This doesn't
work in Clang as Clang doesn't preserve the ordering of inline asm relative to
functions & thus the labels in the DWARF end up wrong.

Clang has no intention of ever supporting this (though I don't have a good
reference on that discussion/conclusion) & so we'll mark this as "untested"
when using Clang. This test doesn't exercise Clang's debug info emission anyway
(since it only contains hardcoded DWARF, rather than using the compiler's debug
output) so it's not going to hide any Clang debug info bugs.


git-svn-id: https://llvm.org/svn/llvm-project@170715 91177308-0d34-0410-b5e6-96231b3b80d8