Teach the extras repository to actually install the clang-format tool.
We should also install the clang-format.py script somewhere that makes
since for a Vim integration script. I don't know where that is though,
so just installing the binary for now. This is enough to let me use the
script from a checkout combined with the installed (and thus less likely
to crash or be slow) clang-format binary.

git-svn-id: https://llvm.org/svn/llvm-project@171164 91177308-0d34-0410-b5e6-96231b3b80d8