[ms-inline asm] Force the use of a base pointer if the MachineFunction includes MS-style inline assembly.
This is a follow-on to r175334.  Forcing a FP to be emitted doesn't ensure it
will be used.  Therefore, force the base pointer as well.  We now treat MS
inline assembly in the same way we treat functions with dynamic stack
realignment and VLAs.  This guarantees the BP will be used to reference 
parameters and locals.
rdar://13218191

git-svn-id: https://llvm.org/svn/llvm-project@175576 91177308-0d34-0410-b5e6-96231b3b80d8