Perform the atomic operations on a global, rather then a local stack variable. This was causing failures on one of our internal -O0 testers.
The issue appears to be that spills/refills to/from the stack are clearing the
monitor and causing the atomic operations to fail, which results in the benchmark
going into an infinite loop.  This seems to only happen at -O0 because the
fast-regalloc is generating a reload in-between the excusive load and exclusive
store to a stack slot adjacent to the local variable x.
rdar://13363219

git-svn-id: https://llvm.org/svn/llvm-project@176903 91177308-0d34-0410-b5e6-96231b3b80d8