Change debugserver to open the socket it listens to in INADDR_LOOPBACK mode by default ("localhost only") instead of INADDR_ANY ("accept connections from any system").
Add a new command line argument to debugserver, --open-connection
or -H which will enable the previous behavior.  It would be used
if you were doing two-system debugging, with lldb running on one
system and debugserver running on the other.  But it is a less
common workflow and should not be the default.

<rdar://problem/12583284> 


git-svn-id: https://llvm.org/svn/llvm-project@177790 91177308-0d34-0410-b5e6-96231b3b80d8